improve damage model

+++ src/ecs/arms/Weapon.h
@@ -10,7 +10,7 @@ public:
 
     virtual int recharge() { return 500; }
 
-    virtual float power() { return 0.0f; }
+    virtual int power() { return 0.0f; }
 
     virtual float throwback() { return 0.0f; }
